Extract MocksAuthFactory trait from LogoutTest
